Separate MachineListControls from MachineList.
