+++ ui/src/app/machines/views/MachineList/MachineListControls/FilterAccordion/FilterAccordion.test.js
