+++ ui/src/app/machines/views/MachineList/MachineListControls/FilterAccordion/_index.scss
