+++ ui/src/app/machines/views/MachineList/MachineListControls/FilterAccordion/index.js
