+++ ui/src/app/machines/views/MachineList/MachineListControls/GroupSelect/GroupSelect.test.js
