+++ ui/src/app/machines/views/MachineList/MachineListControls/GroupSelect/index.js
