+++ ui/src/app/machines/views/MachineList/MachineListControls/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MachineListControls";
